Add validation and saving of game settings.
With the new commit, when a game is created, the body, which includes
the game settings, is correctly checked and saved in the lobby.
Therefore, the lobby class has been added, the newLobbySchema has been
introduced, and a function to validate if only valid modifier
combinations are set has been added. Additionally, some other small
changes have been made.